requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
python-dotenv
pymongo
bson
openai
langchain
langchain-community
chromadb
pdfplumber
PyPDF2
python-docx
docx2txt
rapidfuzz
opencv-python
numpy
pytesseract
python-dateutil
